@@ -1,5 +1,5 @@
 {
-  "name": "platform-web",
+  "name": "Bedrock",
   "version": "0.0.1",
   "license": "MIT",
   "engines": {
@@ -40,6 +40,7 @@
     "koa": "^2.6.2",
     "koa-basic-auth": "^4.0.0",
     "koa-compress": "3.0.0",
+    "koa-mount": "^4.0.0",
     "koa-static": "^5.0.0",
     "koa-webpack": "^5.2.4",
     "less": "2.7.3",
@@ -70,63 +71,17 @@
     "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.12.0",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-plugin-react": "^7.12.3",
-    "lint-staged": "^8.1.0",
-    "prettier": "^1.15.3",
-    "prettier-eslint": "^8.8.2"
-  },
-  "eslintConfig": {
-    "env": {
-      "es6": true,
-      "browser": true,
-      "node": true
-    },
-    "parser": "babel-eslint",
-    "parserOptions": {
-      "ecmaFeatures": {
-        "legacyDecorators": true
-      }
-    },
-    "extends": [
-      "eslint:recommended",
-      "plugin:react/recommended",
-      "plugin:jest/recommended",
-      "prettier"
-    ],
-    "rules": {
-      "react/display-name": 0,
-      "react/prop-types": 0,
-      "no-console": [
-        "warn",
-        {
-          "allow": [
-            "warn",
-            "error",
-            "info"
-          ]
-        }
-      ]
-    },
-    "settings": {
-      "import/resolver": {
-        "node": {
-          "moduleDirectory": [
-            "node_modules",
-            "src"
-          ]
-        }
-      }
-    }
-  },
-  "prettier": {
-    "singleQuote": true,
-    "arrowParens": "always"
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-jest": "^23.8.2",
+    "eslint-plugin-react": "^7.19.0",
+    "lint-staged": "^10.2.2",
+    "prettier": "^2.0.5",
+    "prettier-eslint": "^9.0.1"
   },
   "volta": {
-    "node": "10.15.1",
-    "yarn": "1.13.0"
+    "node": "12.16.3",
+    "yarn": "1.22.4"
   }
 }
