@@ -1,7 +1,5 @@
-import React from 'react';
-import { Label } from 'semantic-ui-react';
 
-function round(value, decimals) {
+export function round(value, decimals) {
   return Number(Math.round(value + 'e' + decimals) + 'e-' + decimals);
 }
 
